+++ packages/common/config/load-config.js
@@ -489,6 +489,9 @@ const loadConfig = async (
       env: "KS_ISLOCAL",
       defaultFunction: (config) => !config.ci,
     },
+    provider: {
+      env: "KS_PROVIDER",
+    },
     clusterEnvironments: {
       transform: (value) => ({
         ...{
@@ -552,7 +555,7 @@ const loadConfig = async (
               `ci-${config.projectName || config.repositoryName}`,
               `${config.projectName || config.repositoryName}-ci`,
             ]
-          : `${config.projectName || config.repositoryName}-ci`,
+          : `ci-${config.projectName || config.repositoryName}`,
       transform: async (value, config) => {
         if (!Array.isArray(value)) {
           return value